Discover [WolverineHandlerModule] assemblies without a filesystem probe (GH-2905)
HandlerDiscovery.DiscoverHandlerModules used AssemblyFinder.FindAssemblies (a
Directory.EnumerateFiles bin-directory probe) to find [WolverineHandlerModule]
assemblies — trim/AOT-hostile and slow at cold start.

- Extract the GH-2902 deployment-list loader (TRUSTED_PLATFORM_ASSEMBLIES, with a
  reference-graph fallback) into a shared ModuleAssemblyLoader, used by both
  extension discovery and handler-module discovery.
- DiscoverHandlerModules now ensures the deployed module assemblies are loaded via
  that shared loader, then picks the [WolverineHandlerModule]-marked ones out of the
  loaded set — no AssemblyFinder filesystem probe. [WolverineHandlerModule] derives
  from [JasperFxAssembly], same eligibility as [WolverineModule].

Behavior is unchanged for the existing IncludeHandlerModules=true test (Module2 is
[assembly: WolverineHandlerModule]); adds a focused handler_module_discovery test
that calls DiscoverHandlerModules directly. Module1 ([WolverineModule] extension) +
Module2 (handler module) both flow through the shared loader.
