+++ src/Wolverine/WolverineOptions.Extensions.cs
@@ -25,7 +25,7 @@ public sealed partial class WolverineOptions
         // deployed [WolverineModule] assemblies so "reference the package and it auto-activates"
         // still works — using the runtime's resolved deployment list, not the old bin-directory
         // glob (Directory.EnumerateFiles) that AssemblyFinder.FindAssemblies used.
-        ensureReferencedModuleAssembliesAreLoaded();
+        Wolverine.Runtime.ModuleAssemblyLoader.EnsureDeployedModuleAssembliesAreLoaded(ApplicationAssembly);
 
         // Include any [WolverineModule]-marked assembly that is now loaded so its handlers
         // participate in discovery, exactly as the old ExtensionLoader.IncludeExtensionAssemblies
@@ -67,139 +67,6 @@ public sealed partial class WolverineOptions
     private static bool isDeclaredModuleExtension(Type type)
     {
         return type == type.Assembly.GetAttribute<WolverineModuleAttribute>()?.WolverineExtensionType;
-    }
-
-    // Framework assembly name prefixes we never need to walk into. No Wolverine extension/module
-    // assembly uses these prefixes, so skipping them keeps the reference-graph walk from loading
-    // the whole BCL closure.
-    private static readonly string[] _nonModulePrefixes =
-        ["System.", "System,", "Microsoft.", "netstandard", "mscorlib", "WindowsBase", "Newtonsoft."];
-
-    // Make sure every deployed [WolverineModule] assembly is loaded so its source-generated manifest
-    // can be read below. This intentionally replaces AssemblyFinder.FindAssemblies'
-    // Directory.EnumerateFiles bin-directory probe.
-    //
-    // It does NOT just walk Assembly.GetReferencedAssemblies(): the C# compiler prunes a referenced
-    // assembly whose types the application never uses directly, so a "reference the package and it
-    // auto-activates" module like WolverineFx.RuntimeCompilation is absent from the metadata
-    // reference graph even though it is deployed. Instead we use the runtime's resolved deployment
-    // list (TRUSTED_PLATFORM_ASSEMBLIES) - not a recursive filesystem scan - which DOES include those
-    // pruned-but-deployed assemblies. The reference-graph walk is kept as a fallback for hosts that
-    // don't surface that list (e.g. single-file deployments).
-    [UnconditionalSuppressMessage("Trimming", "IL2026",
-        Justification = "Loads assemblies from the runtime's resolved deployment list / declared reference graph (Assembly.Load), not a filesystem probe. AOT/trim-published apps reference their extension assemblies statically and register extensions explicitly (ExtensionDiscovery.ManualOnly); see GH-2902 / AOT guide.")]
-    private void ensureReferencedModuleAssembliesAreLoaded()
-    {
-        var seen = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        foreach (var assembly in AppDomain.CurrentDomain.GetAssemblies().Where(a => !a.IsDynamic))
-        {
-            var name = assembly.GetName().Name;
-            if (name != null)
-            {
-                seen.Add(name);
-            }
-        }
-
-        if (!tryLoadFromDeploymentList(seen))
-        {
-            walkReferenceGraph(seen);
-        }
-    }
-
-    // Load the non-framework assemblies the runtime resolved for this app (the deployment closure),
-    // which includes referenced module packages whose types the app doesn't use directly. Returns
-    // false when the host doesn't expose TRUSTED_PLATFORM_ASSEMBLIES so the caller can fall back.
-    private bool tryLoadFromDeploymentList(HashSet<string> seen)
-    {
-        if (AppContext.GetData("TRUSTED_PLATFORM_ASSEMBLIES") is not string list || list.Length == 0)
-        {
-            return false;
-        }
-
-        foreach (var path in list.Split(Path.PathSeparator, StringSplitOptions.RemoveEmptyEntries))
-        {
-            var name = Path.GetFileNameWithoutExtension(path);
-            if (name.IsEmpty() || isNonModuleAssembly(name) || !seen.Add(name))
-            {
-                continue;
-            }
-
-            try
-            {
-                Assembly.Load(new AssemblyName(name));
-            }
-            catch (Exception)
-            {
-                // Not loadable by simple name (native image, resource/satellite assembly, …); skip.
-            }
-        }
-
-        return true;
-    }
-
-    // Fallback: walk the reference graph from the application assembly plus everything already loaded,
-    // loading each non-framework referenced assembly. Misses compiler-pruned references, but better
-    // than nothing when no deployment list is available.
-    [UnconditionalSuppressMessage("Trimming", "IL2026",
-        Justification = "Reference-graph fallback (Assembly.GetReferencedAssemblies/Assembly.Load) used only when the runtime deployment list is unavailable. AOT/trim-published apps register extensions explicitly (ExtensionDiscovery.ManualOnly); see GH-2902 / AOT guide.")]
-    private void walkReferenceGraph(HashSet<string> seen)
-    {
-        var queue = new Queue<Assembly>();
-
-        void enqueue(Assembly assembly)
-        {
-            var name = assembly.GetName().Name;
-            if (name != null && seen.Add(name))
-            {
-                queue.Enqueue(assembly);
-            }
-        }
-
-        foreach (var assembly in AppDomain.CurrentDomain.GetAssemblies().Where(a => !a.IsDynamic).ToArray())
-        {
-            // Already in `seen` from the caller; enqueue for walking without re-adding.
-            queue.Enqueue(assembly);
-        }
-
-        if (ApplicationAssembly != null)
-        {
-            enqueue(ApplicationAssembly);
-        }
-
-        while (queue.Count > 0)
-        {
-            foreach (var reference in queue.Dequeue().GetReferencedAssemblies())
-            {
-                var name = reference.Name;
-                if (name == null || seen.Contains(name) || isNonModuleAssembly(name))
-                {
-                    continue;
-                }
-
-                try
-                {
-                    enqueue(Assembly.Load(reference));
-                }
-                catch (Exception)
-                {
-                    // A reference we can't resolve can't be a discoverable module; ignore it.
-                    seen.Add(name);
-                }
-            }
-        }
-    }
-
-    private static bool isNonModuleAssembly(string name)
-    {
-        foreach (var prefix in _nonModulePrefixes)
-        {
-            if (name.StartsWith(prefix, StringComparison.OrdinalIgnoreCase))
-            {
-                return true;
-            }
-        }
-
-        return false;
     }
 
     // The manifest stores extension types as a DAM-less Type[] (typeof(...) literals emitted by the
